Lab Activity 2 Changes (09/20/23)

+++ EspinaITELEC1C/Models/StudentModel.cs
@@ -13,7 +13,7 @@
         public DateTime DateEnrolled { get; set; }
         public Course StudentCourse { get; set; }
         public string StudentEmail { get; set; }
-        public float GeneralWeightedAverage { get; set; }
+        public double GeneralWeightedAverage { get; set; }
 
     }
 }
